+++ Source/Shaders/ModelExperimental/MaterialStageFS.glsl
@@ -1,16 +1,3 @@
-czm_modelMaterial defaultModelMaterial()
-{
-    czm_modelMaterial material;
-    material.diffuse = vec3(1.0);
-    material.specular = vec3(0.04); // dielectric (non-metal)
-    material.roughness = 0.001;
-    material.normal = vec3(0.0, 0.0, 1.0);
-    material.occlusion = 1.0;
-    material.emissive = vec3(0.0);
-    material.alpha = 1.0;
-    return material;
-}
-
 vec3 SRGBtoLINEAR3(vec3 srgbIn) 
 {
     return pow(srgbIn, vec3(2.2));